backend/requirements.txt
langchain-openai>=0.2.0
langchain-core>=0.3.0
langgraph>=0.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0
PyYAML>=6.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
